Update App.cs
Updated to post the robot structural analysis command instead of the print preview command.

Unfortunately this does not work as intended.
No exceptions are caused and the cmd id for this string is successfully resolved (not null),
however upon investigation, revit does not appear to understand this command as the following
entry is written to the current journal:
   'C 01-Mar-2023 15:25:19.693;  DBG_INFO: CustomCtrl_%CustomCtrl_%CustomCtrl_%Analyze%Structural Analysis%AnalysisAndCodeCheck%RobotStructuralAnalysisLinkdoes not exist.: line 894 of E:\Ship\2022_px64\Source\API\RevitAPIUI\Objects\APIUIApplicationHandwritten.cpp.
likewise with the python version, using PostCommand appears to be a dead end...