Replaces PostCommand with DirectRevitAccess as method of invoking export to Robot
Replaced PostCommand invocation of Robot Structural Analysis with direct API access.

This direct API access does indeed work, but does not meet the minimum viable product as it simply automatically invokes the GUI window on save. This therefore requires the user to still select the export options and confirm, but is a step in the right direction!

Further exploration into how the DirectRevitAccess() operates is required, but this will be difficult as this is not published for end users.

A Revit External Command is added, that acts to store the parameters it was called with, which can then be access from other classes, namely the App.cs, as it is not possible to create these programmatically (provided by Revit invoking the command on when the user selects it). This is a fudge in place for the time being.