More or less minimum viable product of export to robot working on save, though with the following caveats
Only works if the user has selected the Add-ins -> External Tools -> Command RevitToRobotInterfacePlugin first
Only works if the user has already exported the current document to Robot, and Robot is still up and running
The exporting progress window still appears during export
The export is running on the main thread and blocking i.e. the user cannot continue working on the active document
System reflection and decomplication is heavily relied upon to get here. This is far from ideal but the only possible route to a solution given the restraints. Any update to the API by the official developers can result in this completely breaking and requiring a re-write.