Added UI Controls to enable and disable export, optimised and improved reliability
Add ExternalCommands (CommandEnable and CommandDisable) that set and clear a static flag in App  (IsEnabled ), enabling and disabling Handler from doing anything. These are called through UI radio buttons that are added to a ribbon under the Add-ins tab.

RevitToRobotCmdAssembly.GetType("REX.DRevit2Robot.clLog") is created, to resolve null exceptions that are otherwise created when Revit tried to log to it during the operations being performed.

Minor optimisations and refactoring