Export is attempted in a thread
Actions are performed in a new thread (created by Handler) as opposed to in the main thread. This is non-functional, with further work required to determine if it is ever going to be possible (As Revit is single threaded).

The following exceptions are thrown when executed:
Exception thrown: 'Autodesk.Revit.Proxy.Exceptions.InvalidOperationExceptionProxy' in APIUIAPI.dll
Exception thrown: 'Autodesk.Revit.Exceptions.InvalidOperationException' in RevitAPIUI.dll
Exception thrown: 'System.Reflection.TargetInvocationException' in mscorlib.dll
Exception has been thrown by the target of an invocation.

With result:
Result of EvilExportToRobot is Failed